+++ assets/js/custom/cookie-consent.js
@@ -1,7 +1,7 @@
 // EJO Cookie Consent
 // Inspired by the Ilmenite Cookie Banner of Bernskiold Media
 
-var ejoCookieConsent = (function ($) {
+var ejoCookieConsent = (function () {
 
     /**
      * Module pattern
@@ -15,16 +15,25 @@ var ejoCookieConsent = (function ($) {
 
     /**
      * Note: erasing cookie with javascript has no use because this script is not
-     *       added to the site when the cookie is present
+     *       added to the site when the cookie is present. 
+     *       Remark: script is added when SCRIPT_DEBUG is set in wp-config.
      */
 
     // Variables
-    priv.debugMode            = false;                          // Debug Mode: true will disable the cookie, allowing you to debug the banner.
-    priv.consentDuration      = 30;                            // Duration in Days: The number of days before the cookie should expire.
-    priv.containerID          = 'cookie-consent-block';        // The ID of the notice container div
-    priv.containerButtonClass = 'close-cookie-consent-block';  // The Class of the accept-button
-    priv.cookieName           = ejoccLocalization.cookieName;  // The name of the cookie
-    priv.cookieActiveValue    = '1';                           // The active value of the cookie.
+    
+    // Debug. It does set the cookie, but the banner is always visible
+    priv.debugMode            = false; 
+    priv.debugMode            = ejoCookieConsentPlugin.debugMode;
+
+    // Duration in Days: The number of days before the cookie should expire.
+    priv.consentDuration      = ejoCookieConsentPlugin.consentDuration; 
+
+    // Self-explanatory
+    priv.cookieName           = ejoCookieConsentPlugin.cookieName; 
+    priv.consentText          = ejoCookieConsentPlugin.consentText;
+    priv.consentButtonText    = ejoCookieConsentPlugin.consentButtonText;
+    priv.root                 = document.querySelector('body');
+    priv.cookieConsentBlock   = undefined;
 
     // Get Cookie helper
     priv.getCookie = function(name) {
@@ -40,66 +49,106 @@ var ejoCookieConsent = (function ($) {
 
     // Set the cookie
     priv.setCookie = function() {
+        var dateToday, value, expiration;        
 
-        // If no debug mode, set the cookie
-        if ( ! priv.debugMode ) {
+        // Set the consent duration into a cookie date string
+        dateToday = new Date();
+        
+        // Create date "dd-mm-yyyy"
+        var dd = dateToday.getDate();
+        var mm = dateToday.getMonth()+1; //January is 0!
+        var yyyy = dateToday.getFullYear();
 
-            // Set the consent duration into a cookie date string
-            var date = new Date();
-            date.setTime(date.getTime()+(priv.consentDuration*24*60*60*1000));
+        // Leading zeroes
+        if(dd<10) { dd = '0'+dd } 
+        if(mm<10) { mm = '0'+mm }
 
-            // Set the actual cookie
-            document.cookie = priv.cookieName + '=' + priv.cookieActiveValue + '; expires=' + date.toGMTString() + '; path=/';
-        }
+        // Set curent date as value
+        value = dd + "-" + mm + "-" + yyyy;
+
+        // Set expiration
+        dateToday.setTime(dateToday.getTime()+(priv.consentDuration*24*60*60*1000));
+        expiration = dateToday.toUTCString();
+
+        // Set the actual cookie
+        document.cookie = priv.cookieName + '=' + value + '; expires=' + expiration + '; path=/';
     };
 
     // Check if cookie consent
     priv.cookieConsentIsGiven = function(argument) {
-        console.log('cookieConsentIsGiven');
-        return (priv.getCookie(priv.cookieName) == priv.cookieActiveValue)
+
+        return ( priv.getCookie(priv.cookieName) );
     };
 
-    priv.onShowCookieConsentBlock = function() {
-        // Get the height of the consent block
-        var consentBlockHeight = $('#' + priv.containerID).innerHeight();
+    // Show 
+    priv.displayCookieConsentBlock = function() {
+        
+        priv.cookieConsentBlock = document.createElement( 'div' );
+        priv.cookieConsentBlock.classList.add('ejo-cookie-consent');
 
-        // Add class to body
-        $('body').addClass('has-cookie-banner');
-        $('body').css('padding-top', consentBlockHeight + 'px');
+        var cookieConsentBlockInner = document.createElement( 'div' );
+        cookieConsentBlockInner.classList.add('ejo-cookie-consent__inner');
+
+        var cookieConsentBlockText = document.createElement( 'div' );
+        cookieConsentBlockText.classList.add('ejo-cookie-consent__text');
+        cookieConsentBlockText.innerHTML = priv.consentText;
+
+        var cookieConsentBlockButton = document.createElement( 'button' );
+        cookieConsentBlockButton.classList.add('ejo-cookie-consent__button');
+        cookieConsentBlockButton.innerHTML = priv.consentButtonText;
+
+        // Add to DOM
+        cookieConsentBlockInner.appendChild(cookieConsentBlockText);
+        cookieConsentBlockInner.appendChild(cookieConsentBlockButton);
+        priv.cookieConsentBlock.appendChild(cookieConsentBlockInner);
+        priv.root.appendChild(priv.cookieConsentBlock);
+
+        // Let body know it has a cookie banner
+        priv.root.classList.add('has-cookie-banner');
+        
     };
 
     priv.hideCookieConsentBlock = function(){
 
-        // Hide the cookie banner
-        $( '#' + priv.containerID).slideToggle(function() {
-            // Remove cookie banner class
-            $('body').removeClass('has-cookie-banner');
-            $('body').css('padding-top', '0px');
-        });
+        // Let body know it doens't have a cookie banner anymore
+        priv.root.classList.remove('has-cookie-banner');
+
+        // Remove
+        // priv.root.removeChild( priv.root.querySelector( '.ejo-cookie-consent' ) );
+
+        // Move the block out of sight
+        var height = priv.cookieConsentBlock.offsetHeight + 'px';
+        priv.cookieConsentBlock.style.marginBottom = '-' + height;
+
+        // Prevent shadow from being shown
+        priv.cookieConsentBlock.style.boxShadow = 'none';
 
         // Set the cookie
         priv.setCookie();
     };
 
-    publ.init = function() {
+    priv.init = function() {
 
         // Do stuff if no consent for cookie yet
-        if( ! priv.cookieConsentIsGiven() ){
-            priv.onShowCookieConsentBlock();
-        }
+        if ( ! priv.cookieConsentIsGiven() || priv.debugMode ) {
+            
+            // Show consent block
+            priv.displayCookieConsentBlock();
 
-        // Add event handler to close button
-        $( '.' + priv.containerButtonClass).click(priv.hideCookieConsentBlock);
+            // Listen to button
+            priv.cookieConsentBlock.querySelector( '.ejo-cookie-consent__button' ).addEventListener( 'click', priv.hideCookieConsentBlock );
+        }
     };
 
     /**
      * Share public properties and methods with the global namespace
      */
-    return publ;
+    return {
+        'init': priv.init
+    };
 
-})(jQuery);
+})();
 
-
-jQuery(document).ready(function($) {
+document.addEventListener("DOMContentLoaded", function() {
     ejoCookieConsent.init();
 });